update headers...forgot them to commit since a few revisions

+++ zipscript/include/stats.h
@@ -1,7 +1,7 @@
 #ifndef _STATS_H_
 #define _STATS_H_
 
-void updatestats(struct VARS *, struct USERINFO **, struct GROUPINFO **, char *, char *, unsigned int, int, int);
+void updatestats(struct VARS *, struct USERINFO **, struct GROUPINFO **, char *, char *, off_t, unsigned int, unsigned int);
 void sortstats(struct VARS *, struct USERINFO **, struct GROUPINFO **);
 void showstats(struct VARS *, struct USERINFO **, struct GROUPINFO **);
 void get_stats(struct VARS *, struct USERINFO **);
